字符串拼接和String.intern()方法

+++ jvmcsharp/instructions/stores/Xastore.cs
@@ -27,7 +27,7 @@ namespace jvmcsharp.instructions.stores
         public override void Execute(Frame frame)
         {
             var stack = frame.OperandStack;
-            var val = stack.Pop<sbyte>();
+            var val = stack.Pop<int>();
             var index = stack.Pop<int>();
             var arrRef = stack.Pop<ArrayObject>()
                 ?? throw new Exception("java.lang.NullPointException");
@@ -36,7 +36,7 @@ namespace jvmcsharp.instructions.stores
             {
                 throw new Exception("ArrayIndexOutOfBoundsException");
             }
-            vals[index] = val;
+            vals[index] = (sbyte)val;
         }
     }
